% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-funcs.R
\name{list_parallel.lsm}
\alias{list_parallel.lsm}
\title{List Available Parallelized \code{landscapemetrics} Functions}
\usage{
list_parallel.lsm()
}
\value{
A vector.
}
\description{
Get a vector with names of functions available in \code{parallel.lsm}.
Functions are created when the package is loaded and will differ based on
what version of \code{landscapemetrics} is used. See \code{\link[=parallel.lsm]{parallel.lsm()}}
for arguments specific to the parallel versions of these functions.
}
