% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-lsm.R
\name{parallel.lsm}
\alias{parallel.lsm}
\title{Run \code{landscapemetrics} Functions in Parallel}
\usage{
parallel.lsm(..., func, split_on = NULL, join_tbl = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{func}.}

\item{func}{function. A function from the \link[landscapemetrics:landscapemetrics]{landscapemetrics} package.}

\item{split_on}{character. Names of arguments in \code{...} to split between workers.
By default, the name of the first argument in \code{...} is used.
Multiple arguments can be specified by using a vector. Arguments specified here
should be split into groups (such as by using \code{\link[base:split]{base::split()}} or \code{\link[terra:split]{terra::split()}})
that equal the number of workers you want to use. The number of available workers
can be checked using \code{\link[parallelly:availableCores]{parallelly::availableCores()}}. It is generally recommended to
use at most n-1 or n-2 cores compared to the total available.}

\item{join_tbl}{logical. If the output should be joined into a single table. Outputs
are only combined if the result for all workers is a data frame or similar object
(i.e., tibble). If \code{FALSE}, a list with the results for each worker is returned instead.}
}
\value{
format varies based on \code{func}, typically a tibble.
}
\description{
Run \code{landscapemetrics} functions across multiple workers at the same time (in parallel).
\code{\link[=parallel.lsm]{parallel.lsm()}} can be used to call any function from \code{landscapemetrics} as a parallel process.
Other parallel functions are functionally identical but with the \code{func} argument already specified
internally. All functions are formatted as: \verb{parallel.*()}.
}
